chore(precompiles): remove permit from tip20 and linking_usd (#580)
We don't need permit because of default account abstraction, and tokens
can always fallback to permit2.
Since permit is not available in all existing ERC20 tokens, almost all
DApps are built with the assumption that permit will not be present in a
token, so this shouldn't break compatibility with anything.

This also removes a decent chunk of complexity from tip20 tokens.