fix(tempo-bench): clear txpool for all URLs and get start block from sent txs (#1123)
We run `tempo-bench` on an RPC that's load balancing the requests
between multiple nodes, so we need to pass them all as arguments to
`target_urls` and clear the tx pool for each.

Additionally, this fixes the way we query the start block number for the
report: we can't just query for the latest block when we start sending
txs, because it most likely be still too early and have no transactions
in it. For example, see
https://bench.tempoxyz.dev/report_manual-2025-11-28-13-43-15.html

<img width="702" height="463" alt="image"
src="https://github.com/user-attachments/assets/030a6b8a-6f5d-4fe8-8496-4ab62d3dbec4"
/>