chore: port docs into tempo repo (#1148)
This allows us to atomically commit to both specs and the node
implementations, which makes it easier to keep them in sync.

I am assuming we don't care about commit history here, so chose to just
create a new subdirectory for this. Dealing with git submodules would
have been annoying in the future.

---------

Co-authored-by: Jennifer <jenpaff0@gmail.com>