fix: await FCU on proposal (#1181)
When proposing, reth expects the `parent_hash` to be part of the
canonical chain. However, right now FCU with the parent races the
`send_new_payload` call which results in it failing when engine is not
able to process FCU soon enough.

It seems that issue is responsible for majority of nullifications
happening on Andantino (there's one every few minutes)

Adds a `oneshot` channel that can be used to await the FCU when calling
`canonicalize` and uses it in `handle_propose`