fix: send proposal parent via a newPayload (#1410)
Changes proposal logic to start with sending parent we need to propose
on top of as `newPayload` to reth.

In most cases this will be cheap enough because this block is already
known to reth but in some other cases this might prevent failing
proposal if we do not have this block in the engine yet.

This is still a bit wasteful and not entirely robust because we might be
missing multiple notarized blocks (e.g a parent's parent as well).
Proper fix for this is to listen to notarizations directly so that no
notarized blocks go past reth