fix: multicall3 error, and add sync tests against ethereum mainnet (#1629)
Currently the `multicall` bytecode is deployed at the `multicall3`
address.

This PR also adds tests which makes sure that the bytecode hash and
storage root of these predeployed genesis contracts is always synced
with ethereum mainnet.

The tests have also caught an inconsistency between the permit2
codehashes. Investigating what's the reason, and accordingly we might
need to update it as well.